Moved most scripts from assets to package

+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Logger.cs
@@ -20,7 +20,7 @@ namespace Mediapipe
 
 	public static class Logger
 	{
-		public enum LogLevel : int
+		public enum LogLevel : byte
 		{
 			Fatal,
 			Error,
